Create isHappyTicket
A "lucky" ticket is called a ticket with a number in which the sum of the first half of the digits is equal to the sum of the second half of the digits. The numbers can be of any length, with the only condition that the number of digits is always even.

The ticket with the number 385916 is a lucky one, since 3 + 8 + 5 === 9 + 1 + 6. The ticket with the number 231002 is not a happy one, since 2 + 3 + 1 !== 0 + 0 + 2.

The function returns true if the ticket is lucky, or false if it is not.

isHappyTicket('385916'); // true
isHappyTicket('231002'); // false